refactor(balancing): align deep-module ownership
